finish devices page implementation

+++ application/src/app/pages/devices/devices.component.ts
@@ -8,17 +8,19 @@ import { CardComponent } from "../../shared/card/card.component";
 import { NgClass } from '@angular/common';
 import { MatIconModule } from '@angular/material/icon';
 import { DevicesAddComponent } from "./components/devices-add/devices-add.component";
+import { DevicesEditComponent } from "./components/devices-edit/devices-edit.component";
 
 @Component({
   selector: 'app-devices',
   standalone: true,
-  imports: [CardComponent, NgClass, MatIconModule, DevicesAddComponent],
+  imports: [CardComponent, NgClass, MatIconModule, DevicesAddComponent, DevicesEditComponent],
   templateUrl: './devices.component.html',
   styleUrl: './devices.component.css'
 })
 export class DevicesComponent implements OnInit, OnDestroy {
   private destroyed$ = new Subject<void>();
   public devicesData: {device: Device, incidents: Incident[]}[] = [];
+  public editDevice: Device | undefined = undefined;
 
   constructor(
     private deviceService: DevicesService,
@@ -38,6 +40,15 @@ export class DevicesComponent implements OnInit, OnDestroy {
         this.devicesData.push({device, incidents: deviceIncidents});
       })
     });
+
+    this.deviceService.selectedDeviceID$.pipe(
+      takeUntil(this.destroyed$)
+    ).subscribe((deviceID) => {
+      this.editDevice = this.deviceService.devices.find(d => d.id === deviceID);
+      if (this.editDevice !== undefined) {
+        this.editDevice = {...this.editDevice}
+      }
+    });
   }
 
   ngOnDestroy(): void {
@@ -50,11 +61,26 @@ export class DevicesComponent implements OnInit, OnDestroy {
   }
 
   deleteDevice(deviceID: string) {
-    this.deviceService.deleteDevice(deviceID);
+    if (confirm(`You are about to delete the device ${this.deviceService.devices.find(d => d.id === deviceID)!.name}`)) {
+      this.deviceService.deleteDevice(deviceID);
+    }
   }
 
   connectDevice(deviceID: string) {
     this.deviceService.connectToDevice(deviceID);
+  }
+
+  selectDevice(deviceID: string) {
+    this.deviceService.selectedDeviceID = deviceID;
+  }
+
+  closeEdit() {
+    this.deviceService.selectedDeviceID = undefined;
+  }
+
+  edit(device: Device) {
+    this.deviceService.updateDevice({...device});
+    this.deviceService.selectedDeviceID = undefined;
   }
 
 }
